Updating DevTools component
