+++ src/components/DevTools/DevTools.styles.js
@@ -1,0 +1,9 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const DevToolsStyles = makeStyles(theme => ({
+  toolsIcon: {
+    fontSize: theme.pxToRem(32)
+  }
+}));
+
+export default DevToolsStyles;
