+++ src/components/DevTools/index.js
@@ -1,0 +1,3 @@
+import DevTools from './DevTools';
+
+export default DevTools;
